# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Competency Platform REST API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /competencies:
    get:
      operationId: listCompetencies
      tags: [competencies]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
        - { name: take, in: query, schema: { type: integer } }
        - { name: skip, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Competencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competency"
    post:
      operationId: createCompetency
      tags: [competencies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetencyInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competency"
...
  /competencies/{id}:
    put:
      operationId: updateCompetency
      tags: [competencies]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetencyInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competency"
    delete:
      operationId: deleteCompetency
      tags: [competencies]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /competencies/import:
    post:
      operationId: importCompetencies
      tags: [competencies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetencyImportInput"
      responses:
        "200":
          description: Imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      parameters:
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    put:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    ImportResult:
      type: object
      properties:
        imported: { type: integer }
      required: [imported]
...
    Competency:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        level: { type: string }
        description: { type: ["string", "null"] }
        indicators: { type: ["string", "null"] }
        isShared: { type: boolean }
        sharedKey: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, name, type, level, isShared]
...
    CompetencyInput:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        level: { type: string }
        description: { type: string }
        indicators: { type: string }
      required: [name, type]
...
    CompetencyImportInput:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/CompetencyInput"
      required: [rows]
...
    Employee:
      type: object
      properties:
        id: { type: string }
        employeeNumber: { type: string }
        name: { type: string }
        jobId: { type: ["string", "null"] }
        departmentId: { type: ["string", "null"] }
        gradeId: { type: ["string", "null"] }
        managerId: { type: ["string", "null"] }
        extraFields:
          type: object
          additionalProperties: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
        jobName: { type: ["string", "null"] }
        departmentName: { type: ["string", "null"] }
        gradeName: { type: ["string", "null"] }
        managerName: { type: ["string", "null"] }
      required: [id, employeeNumber, name]
    EmployeeInput:
      type: object
      properties:
        name: { type: string }
        employeeNumber: { type: string }
        jobId: { type: ["string", "null"] }
        departmentId: { type: ["string", "null"] }
        gradeId: { type: ["string", "null"] }
        managerId: { type: ["string", "null"] }
        extraFields:
          type: object
          additionalProperties: { type: string }
      required: [name, employeeNumber]
...